+++ manualinput.js
@@ -1,0 +1,21 @@
+class ManualInput extends InputModule {
+	constructor() {
+		super();
+	}
+
+	
+
+// function keyPressed() {
+// 		console.log("key pressed");
+// 		if (keyCode === UP_ARROW) {
+// 			console.log("up received from p5");
+// 			ManualInput.up();
+// 		} else if (keyCode === DOWN_ARROW) {
+// 			ManualInput.down();
+// 		} else if (keyCode === LEFT_ARROW) {
+// 			ManualInput.left();
+// 		} else if (keyCode === RIGHT_ARROW) {
+// 			ManualInput.right();
+// 		}
+// 	}
+}